image changed to legacy and future image to image
